Atlas theme - correction of arrows position
Corrected the position of the arrows in the default timeline 1px to the
right so that they are not overlapped by their panels.

Created a button to scroll to top of the page

For this button, I added
a new page to the componets section because it
would conflict with other
elements in the buttons page and  because I
think this item is more a
navigational aid than a real button that
triggers some kind of action.